oop1-2103/2025-1

+++ RoomDirect/Room.cpp
@@ -1,7 +1,7 @@
 #include "Room.h"
 
 
-Room::Room(int roomNum, int price) {
+Room::Room(int roomNum, double price) {
     this->roomNum = roomNum;
     this->price = price;
 }
@@ -23,12 +23,13 @@ int Room::getNumOfroom() {
 void Room::setIsBooked(bool isBooked) {
     this->isBooked = isBooked;
 }
-void Room::setPrice(int price) {
+void Room::setPrice(double price) {
     this->price = price;
 }
 void Room::setNumOfclient(int numOfclient) {
     this->numOfclient = numOfclient;
 }
+
