+++ RoomDirect/Standart.cpp
@@ -1,6 +1,6 @@
 #include "Standart.h"
 
-Standart::Standart(int roomNum, int price):Room(roomNum,price)
+Standart::Standart(int roomNum, double price):Room(roomNum,price)
 {
 
 }
