+++ RoomDirect/VIP.cpp
@@ -1,6 +1,6 @@
 #include "VIP.h"
 
-VIP::VIP(int roomNum, int price):Room(roomNum,price)
+VIP::VIP(int roomNum, double price):Room(roomNum,price)
 {
 
 }
