+++ RoomDirect/VIP.h
@@ -5,7 +5,7 @@
 
 class VIP: public Room {
 public:
-    VIP(int roomNum,int price);
+    VIP(int roomNum,double price);
 
 private:
